backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-multipart
google-generativeai

